requirements.txt
opencv-python~=4.8.0.76

numpy~=1.25.2
Flask~=2.3.2
pyserial~=3.5
Pillow~=10.0.0